requirements.txt
# Core ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Model persistence
joblib==1.3.2

# Data visualization and reporting
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Data validation
great-expectations==0.17.15

# MLOps tools
dvc==3.15.0
dvclive==2.11.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
